@@ -1,23 +1,24 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.example.selenium</groupId>
     <artifactId>selenium-springboot-example</artifactId>
-    <version>1.0.0</version>
+    <version>1.1.3</version>
 
     <properties>
 
-        <java.version>11</java.version>
+        <java.version>18</java.version>
         <encoding>UTF-8</encoding>
         <project.build.sourceEncoding>${encoding}</project.build.sourceEncoding>
         <project.reporting.outputEncoding>${encoding}</project.reporting.outputEncoding>
 
-        <springboot.version>2.7.1</springboot.version>
-        <webdriver.version>5.2.1</webdriver.version>
-        <cucumber.version>7.4.1</cucumber.version>
-        <selenium.version>4.3.0</selenium.version>
+        <maven.compiler.version>3.11.0</maven.compiler.version>
+        <springboot.version>3.1.2</springboot.version>
+        <webdriver.version>5.4.1</webdriver.version>
+        <cucumber.version>7.13.0</cucumber.version>
+        <selenium.version>4.11.0</selenium.version>
 
     </properties>
 
@@ -26,9 +27,10 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
+                <version>${maven.compiler.version}</version>
                 <configuration>
-                    <source>8</source>
-                    <target>8</target>
+                    <source>18</source>
+                    <target>18</target>
                 </configuration>
             </plugin>
         </plugins>
